refactor capability class - organize functionality
